Add dokku-link LINK file
On dokku, this is used by the dokku-link plugin to link the
ElasticSearch container to the app containers. This lets us leave
sane defaults, such as localhost:9200, in production.ini, but take
advantage of f70c6ea (getting ElasticSearch config from environment)
so that deploys to dokku can happen straight from development branches.